# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AI Content Factory API
  version: 0.1.0
paths:
...
  /api/v1/external-automation/bindings/{bindingId}/rotate-callback-token:
    post:
      summary: Rotate callback token
      tags: [external-automation]
      operationId: rotateCallbackToken
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - name: bindingId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Callback token rotated }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
      security:
        - bearerAuth: []
  /api/v1/external-automation/bindings/{bindingId}/callback-auth:
    patch:
      summary: Update callback auth
      tags: [external-automation]
      operationId: updateCallbackAuth
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - name: bindingId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Callback auth updated }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
      security:
        - bearerAuth: []
...
  /api/v1/external-automation/callback-logs:
    get:
      summary: List callback logs
      tags: [external-automation]
      operationId: listCallbackLogs
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200': { description: Callback log list }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
      security:
        - bearerAuth: []
...
components:
  parameters:
    RequestIdHeader:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
    PageSize:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Authenticated caller is not allowed to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
        reason:
          type: string
      required: [reason]
    APIError:
      type: object
      properties:
        code:
          type: string
          enum: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, IDEMPOTENCY_CONFLICT, WORKFLOW_RUN_FAILED, AGENT_OUTPUT_INVALID, LLM_PROVIDER_ERROR, INTERNAL_ERROR]
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
      required: [code, message, details]
    ErrorEnvelope:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        data:
          nullable: true
        error:
          $ref: '#/components/schemas/APIError'
        request_id:
          type: string
      required: [success, data, error, request_id]
...